Apply review comment
From: https://github.com/jjerphan/scikit-learn/pull/14#pullrequestreview-1004501769
